+++ CHANGELOG.md
@@ -4,8 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 2021-01-03 - Edu
+## Added
+- **utils/heredoc_handler.c** Added with functions to handle all heredocs and unclosed quotes
+ at start of processline
+
 ## 2021-01-02 - Edu
-## Added 
+## Added
 - **utils/token_handler2.c** Added **eval_token** function to returns if token is PIPE, REDIR, CMD...
 ## Changed
 - Replaced **strcmp** & **modstrcmp** by **libft 's ft_strcmp** (ft_modstrcmp.c to old)